refactor: replace APIError type
