+++ packages/client/src/api/typings.ts
@@ -1,10 +1,6 @@
 import { type LEADERBOARD_TEAM_NAME } from '../config/constants';
 import { type SortOption } from '../store/features/leaderboard/typings';
 
-export type APIError = {
-  reason: string;
-};
-
 export type LeaderboardRecordData = {
   username: string;
   score: number;
