+++ packages/client/src/utils/apiUtils.ts
@@ -1,4 +1,4 @@
-import { type APIError } from '../api/typings';
+import { type APIError } from '../app.typings';
 
 export const apiHasError = (response: Record<string, unknown> | APIError): response is APIError => {
   return typeof response === 'object' && 'reason' in response;
